libs: reduce one module level

a
