+++ lua/editor/keymap.lua
@@ -1,44 +1,8 @@
--- map create a new mapping
----@param mode string|table specify vim mode
----@param lhs string specify the new keymap
----@param rhs string|function specify the keymap or commands
----@param opts table|nil setting options. Default: { noremap = true, silent = true, expr = false }
-local function map(mode, lhs, rhs, opts)
-  local options = {
-    noremap = true,
-    silent = true,
-    expr = false,
-  }
-  if opts then
-    options = vim.tbl_extend("force", options, opts)
-  end
-  if type(rhs) == "function" then
-    options.callback = rhs
-    rhs = ""
-  end
-  local stat, error = pcall(vim.keymap.set, mode, lhs, rhs, options)
-  if not stat then
-    vim.notify(error, vim.log.levels.ERROR, {
-      title = "keymap",
-    })
-  end
-end
-
-local function nmap(...)
-  map("n", ...)
-end
-
-local function xmap(...)
-  map("x", ...)
-end
-
-local function imap(...)
-  map("i", ...)
-end
-
-local d = function(s)
-  return { desc = s }
-end
+local nmap = require("libs.keymaps").nmap
+local xmap = require("libs.keymaps").xmap
+local imap = require("libs.keymaps").imap
+local map = require("libs.keymaps").map
+local d = require("libs.keymaps").d
 
 vim.g.mapleader = ";"
 
@@ -89,7 +53,7 @@ nmap("<ESC>", ":nohlsearch<CR>", d("Close search highlight"))
 map("i", "<A-;>", "<ESC>", d("Exit the insert mode"))
 
 -- kill buffer with ;q , quit window with :q.
-nmap(";q", require("plugins.libs.bufdel").delete_buffer)
+nmap(";q", require("libs.bufdel").delete_buffer)
 
 -- Write and quit. Alias for :wq<CR>
 nmap(";x", ":x<CR>")
@@ -238,6 +202,13 @@ nmap("=", _cmd_sequence("increment", "normal"))
 nmap("-", _cmd_sequence("decrement", "normal"))
 map("v", "=", _cmd_sequence("increment", "visual"))
 map("v", "-", _cmd_sequence("decrement", "visual"))
+
+--
+-- Git
+--
+nmap("gic", ":Git commit -sS<CR>")
+nmap("giP", ":Dispatch! git push ", { silent = false })
+nmap("<leader>g", [[<CMD>vertical Git<CR>]])
 
 return {
   map = map,
